requirements.txt
tensorflow-gpu==2.0.1
opencv-python==3.4.3.18
PyQt5
numpy